requirements.txt
mysql-connector-python==8.0.21
numpy==1.19.0
opencv-contrib-python==4.3.0.36
Pillow==7.2.0
protobuf==3.12.2
PyQt5==5.15.0
PyQt5-sip==12.8.0
six==1.15.0
